@@ -17,7 +17,7 @@
         "psr-4": { "Ecommit\\FrequencyGenerator\\Tests\\": "tests/" }
     },
     "require": {
-        "php": "^7.2|^8.0"
+        "php": "^7.4|^8.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
